(BKR-358) extract helper methods to test_helper.rb
